refactor(search): make searchbox url env var optional

+++ server/src/components/search/elastic.ts
@@ -1,16 +1,23 @@
 import { Client } from '@elastic/elasticsearch';
+import { RequestNDBody } from '@elastic/elasticsearch/lib/Transport';
 
 import { searchConfig } from '../../config';
 
-export const client = new Client({ node: searchConfig.host });
+export const client = searchConfig.host
+  ? new Client({ node: searchConfig.host })
+  : null;
 
 export const indexExists = async (index: string): Promise<boolean> => {
+  if (client == null) return false;
+
   // eslint-disable-next-line security/detect-non-literal-fs-filename
   const { body } = await client.indices.exists({ index });
   return Boolean(body);
 };
 
 export const refreshIndex = async (index: string): Promise<void> => {
+  if (client == null) return;
+
   await client.indices.refresh({ index });
 };
 
@@ -18,11 +25,26 @@ export const createIndex = async <TMappings = any>(
   index: string,
   mappings: TMappings,
 ): Promise<void> => {
+  if (client == null) return;
+
   await client.indices.create({
     index,
     body: {
       mappings,
     },
+  });
+};
+
+export const bulkIndex = async ({
+  body,
+}: {
+  body: RequestNDBody;
+}): Promise<void> => {
+  if (client == null) return;
+
+  await client.bulk({
+    refresh: true,
+    body,
   });
 };
 
@@ -47,6 +69,8 @@ export const searchIndex = async <TSource = any>({
     _source: TSource;
   }[]
 > => {
+  if (client == null) return [];
+
   const { body } = await client.search({
     index,
     body: {
@@ -64,6 +88,8 @@ export const createDocument = async <T>(
   index: string,
   body: T,
 ): Promise<void> => {
+  if (client == null) return;
+
   await client.index({
     index,
     body,
@@ -76,6 +102,8 @@ export const updateDocument = async <TBody, TQuery = any>(
   query: TQuery,
   body: TBody,
 ): Promise<void> => {
+  if (client == null) return;
+
   const [doc] = await searchIndex({ index, query });
   if (doc) {
     await client.update({
@@ -92,6 +120,8 @@ export const deleteDocument = async <TQuery = any>(
   index: string,
   query: TQuery,
 ): Promise<void> => {
+  if (client == null) return;
+
   const [doc] = await searchIndex({ index, query });
   if (doc) {
     await client.delete({
